Split MCP OAuth helpers for file size gate
