+++ crates/tandem-runtime/src/mcp_parts/part01.rs
@@ -1954,72 +1954,3 @@ fn resolve_secret_header_values(
     }
     out
 }
-
-fn delete_secret_header_refs(
-    secret_headers: &HashMap<String, McpSecretRef>,
-    current_tenant: &TenantContext,
-) {
-    for secret_ref in secret_headers.values() {
-        if let McpSecretRef::Store {
-            secret_id,
-            tenant_context,
-        } = secret_ref
-        {
-            if tenant_context != current_tenant {
-                continue;
-            }
-            let _ = tandem_core::delete_provider_auth_for_tenant(current_tenant, secret_id);
-        }
-    }
-}
-
-fn delete_oauth_secret_ref(oauth: Option<&McpOAuthConfig>, current_tenant: &TenantContext) {
-    let Some(secret_ref) = oauth.and_then(|oauth| oauth.client_secret_ref.as_ref()) else {
-        return;
-    };
-    if let McpSecretRef::Store {
-        secret_id,
-        tenant_context,
-    } = secret_ref
-    {
-        if tenant_context == current_tenant {
-            let _ = tandem_core::delete_provider_auth_for_tenant(current_tenant, secret_id);
-        }
-    }
-}
-
-fn delete_oauth_credential(
-    server_name: &str,
-    oauth: Option<&McpOAuthConfig>,
-    current_tenant: &TenantContext,
-    oauth_security_dir: &Path,
-) {
-    let mut provider_ids = Vec::new();
-    if let Some(provider_id) = oauth
-        .map(|oauth| oauth.provider_id.trim())
-        .filter(|provider_id| !provider_id.is_empty())
-    {
-        provider_ids.push(provider_id.to_string());
-    }
-    let canonical_provider_segment = server_name
-        .trim()
-        .replace(|ch: char| !ch.is_ascii_alphanumeric(), "_")
-        .trim_matches('_')
-        .to_ascii_lowercase();
-    let canonical_provider_id = format!("mcp-oauth::{canonical_provider_segment}");
-    if !canonical_provider_segment.is_empty()
-        && !provider_ids
-            .iter()
-            .any(|provider_id| provider_id == &canonical_provider_id)
-    {
-        provider_ids.push(canonical_provider_id);
-    }
-    for provider_id in provider_ids {
-        let _ = tandem_core::delete_provider_credential_for_tenant_in_dir(
-            oauth_security_dir,
-            current_tenant,
-            &provider_id,
-        );
-        let _ = tandem_core::delete_provider_credential(&provider_id);
-    }
-}
