+++ crates/tandem-server/src/http/mcp.rs
@@ -14,6 +14,9 @@ use uuid::Uuid;
 
 pub(crate) use super::mcp_run_as::call_mcp_tool_for_tenant_with_audit;
 
+mod mcp_base_url;
+pub(super) use mcp_base_url::{mcp_public_base_url_from_config, mcp_public_base_url_from_env};
+
 const BUILTIN_GITHUB_MCP_SERVER_NAME: &str = "github";
 const BUILTIN_GITHUB_MCP_TRANSPORT_URL: &str = "https://api.githubcopilot.com/mcp/";
 const BUILTIN_TANDEM_DOCS_MCP_SERVER_NAME: &str = "tandem-mcp";
@@ -530,29 +533,6 @@ fn effective_mcp_headers(server: &tandem_runtime::McpServer) -> HashMap<String, 
 }
 fn mcp_uses_oauth(server: &tandem_runtime::McpServer) -> bool {
     server.auth_kind.trim().eq_ignore_ascii_case("oauth")
-}
-pub(super) fn mcp_public_base_url_from_config(cfg: &Value) -> Option<String> {
-    cfg.get("hosted")
-        .and_then(Value::as_object)
-        .and_then(|hosted| hosted.get("public_url"))
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(|value| value.trim_end_matches('/').to_string())
-}
-pub(super) fn mcp_public_base_url_from_env() -> Option<String> {
-    [
-        "TANDEM_CONTROL_PANEL_PUBLIC_URL",
-        "HOSTED_CONTROL_PANEL_PUBLIC_URL",
-        "HOSTED_PUBLIC_URL",
-    ]
-    .into_iter()
-    .find_map(|key| {
-        std::env::var(key)
-            .ok()
-            .map(|value| value.trim().trim_end_matches('/').to_string())
-            .filter(|value| !value.is_empty())
-    })
 }
 fn mcp_public_base_url(state: &AppState, cfg: &Value) -> String {
     mcp_public_base_url_from_config(cfg)
